requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.10.1
uuid==1.30 